+++ creme/imblearn/__init__.py
@@ -1,7 +1,11 @@
 """Imbalanced learning."""
+from .random import RandomOverSampler
 from .random import RandomUnderSampler
+from .random import RandomSampler
 
 
 __all__ = [
+    'RandomOverSampler',
     'RandomUnderSampler',
+    'RandomSampler'
 ]