+++ creme/imblearn/random.py
@@ -1,15 +1,14 @@
 import collections
-import random
 
-from .. import base
+from . import base
 
 
-class RandomUnderSampler(base.Wrapper, base.Classifier):
+class RandomUnderSampler(base.Sampler):
     """Random under-sampling.
 
     This is a wrapper for classifiers. It will train the provided classifier by under-sampling the
-    stream of given observations so that the class distribution seen by the classifier ressembles
-    a given desired distribution. This is in fact a discrete version of rejection sampling.
+    stream of given observations so that the class distribution seen by the classifier follows
+    a given desired distribution. The implementation is a discrete version of rejection sampling.
 
     Parameters:
         classifier (base.Classifier)
@@ -25,16 +24,10 @@ class RandomUnderSampler(base.Wrapper, base.Classifier):
     """
 
     def __init__(self, classifier, desired_dist, seed=None):
-        self.classifier = classifier
+        super().__init__(classifier=classifier, seed=seed)
         self.desired_dist = desired_dist
-        self.seed = seed
-        self._rng = random.Random(seed)
         self._actual_dist = collections.Counter()
         self._pivot = None
-
-    @property
-    def _model(self):
-        return self.classifier
 
     def fit_one(self, x, y):
 
@@ -44,22 +37,100 @@ class RandomUnderSampler(base.Wrapper, base.Classifier):
 
         # Check if the pivot needs to be changed
         if y != self._pivot:
-            self._pivot = max(g.keys(), key=lambda x: f[x] / g[x])
+            self._pivot = max(g.keys(), key=lambda y: f[y] / g[y])
         else:
             self.classifier.fit_one(x, y)
             return self
 
         # Determine the sampling ratio if the class is not the pivot
-        M = f[self._pivot] / g[self._pivot]
+        M = f[self._pivot] / g[self._pivot]  # Likelihood ratio
         ratio = f[y] / (M * g[y])
 
-        if self._rng.random() < ratio:
+        if ratio < 1 and self._rng.random() < ratio:
             self.classifier.fit_one(x, y)
 
         return self
 
-    def predict_proba_one(self, x):
-        return self.classifier.predict_proba_one(x)
 
-    def predict_one(self, x):
-        return self.classifier.predict_one(x)
+class RandomOverSampler(base.Sampler):
+    """Random over-sampling.
+
+    This is a wrapper for classifiers. It will train the provided classifier by over-sampling the
+    stream of given observations so that the class distribution seen by the classifier follows
+    a given desired distribution. The implementation is a discrete version of reverse rejection
+    sampling.
+
+    Parameters:
+        classifier (base.Classifier)
+        desired_dist (dict): The desired class distribution. The keys are the classes whilst the
+            values are the desired class percentages. The values must sum up to 1.
+
+    See :ref:`Working with imbalanced data` for example usage.
+
+    """
+
+    def __init__(self, classifier, desired_dist, seed=None):
+        super().__init__(classifier=classifier, seed=seed)
+        self.desired_dist = desired_dist
+        self._actual_dist = collections.Counter()
+        self._pivot = None
+
+    def fit_one(self, x, y):
+
+        self._actual_dist[y] += 1
+        f = self.desired_dist
+        g = self._actual_dist
+
+        # Check if the pivot needs to be changed
+        if y != self._pivot:
+            self._pivot = max(g.keys(), key=lambda y: g[y] / f[y])
+        else:
+            self.classifier.fit_one(x, y)
+            return self
+
+        M = g[self._pivot] / f[self._pivot]
+        rate = M * f[y] / g[y]
+
+        for _ in range(self._rng.poisson(rate)):
+            self.classifier.fit_one(x, y)
+
+        return self
+
+
+class RandomSampler(base.Sampler):
+    """Random sampling by mixing under-sampling and over-sampling.
+
+    This is a wrapper for classifiers. It will train the provided classifier by both under-sampling
+    and over-sampling the stream of given observations so that the class distribution seen by the
+    classifier follows a given desired distribution.
+
+    Parameters:
+        classifier (base.Classifier)
+        desired_dist (dict): The desired class distribution. The keys are the classes whilst the
+            values are the desired class percentages. The values must sum up to 1.
+        sampling_rate (float): The desired ratio of data to sample.
+
+    See :ref:`Working with imbalanced data` for example usage.
+
+    """
+
+    def __init__(self, classifier, desired_dist, sampling_rate=1., seed=None):
+        super().__init__(classifier=classifier, seed=seed)
+        self.desired_dist = desired_dist
+        self.sampling_rate = sampling_rate
+        self._actual_dist = collections.Counter()
+        self._n = 0
+
+    def fit_one(self, x, y):
+
+        self._actual_dist[y] += 1
+        self._n += 1
+        f = self.desired_dist
+        g = self._actual_dist
+
+        rate = self.sampling_rate * f[y] / (g[y] / self._n)
+
+        for _ in range(self._rng.poisson(rate)):
+            self.classifier.fit_one(x, y)
+
+        return self